+++ src/ui/map.test.ts
@@ -236,92 +236,6 @@ describe('Map', () => {
         expect(fixedLngLat(map.unproject([100, 100]))).toEqual({lng: 0, lat: 0});
     });
 
-    describe('#queryRenderedFeatures', () => {
-
-        test('if no arguments provided', done => {
-            createMap({}, (err, map) => {
-                expect(err).toBeFalsy();
-                const spy = jest.spyOn(map.style, 'queryRenderedFeatures');
-
-                const output = map.queryRenderedFeatures();
-
-                const args = spy.mock.calls[0];
-                expect(args[0]).toBeTruthy();
-                expect(args[1]).toEqual({availableImages: []});
-                expect(output).toEqual([]);
-
-                done();
-            });
-        });
-
-        test('if only "geometry" provided', done => {
-            createMap({}, (err, map) => {
-                expect(err).toBeFalsy();
-                const spy = jest.spyOn(map.style, 'queryRenderedFeatures');
-
-                const output = map.queryRenderedFeatures(map.project(new LngLat(0, 0)));
-
-                const args = spy.mock.calls[0];
-                expect(args[0]).toEqual([{x: 100, y: 100}]); // query geometry
-                expect(args[1]).toEqual({availableImages: []}); // params
-                expect(args[2]).toEqual(map.transform); // transform
-                expect(output).toEqual([]);
-
-                done();
-            });
-        });
-
-        test('if only "params" provided', done => {
-            createMap({}, (err, map) => {
-                expect(err).toBeFalsy();
-                const spy = jest.spyOn(map.style, 'queryRenderedFeatures');
-
-                const output = map.queryRenderedFeatures({filter: ['all']});
-
-                const args = spy.mock.calls[0];
-                expect(args[0]).toBeTruthy();
-                expect(args[1]).toEqual({availableImages: [], filter: ['all']});
-                expect(output).toEqual([]);
-
-                done();
-            });
-        });
-
-        test('if both "geometry" and "params" provided', done => {
-            createMap({}, (err, map) => {
-                expect(err).toBeFalsy();
-                const spy = jest.spyOn(map.style, 'queryRenderedFeatures');
-
-                const output = map.queryRenderedFeatures({filter: ['all']});
-
-                const args = spy.mock.calls[0];
-                expect(args[0]).toBeTruthy();
-                expect(args[1]).toEqual({availableImages: [], filter: ['all']});
-                expect(output).toEqual([]);
-
-                done();
-            });
-        });
-
-        test('if "geometry" with unwrapped coords provided', done => {
-            createMap({}, (err, map) => {
-                expect(err).toBeFalsy();
-                const spy = jest.spyOn(map.style, 'queryRenderedFeatures');
-
-                map.queryRenderedFeatures(map.project(new LngLat(360, 0)));
-
-                expect(spy.mock.calls[0][0]).toEqual([{x: 612, y: 100}]);
-                done();
-            });
-        });
-
-        test('returns an empty array when no style is loaded', () => {
-            const map = createMap({style: undefined});
-            expect(map.queryRenderedFeatures()).toEqual([]);
-        });
-
-    });
-
     test('render stabilizes', done => {
         const style = createStyle();
         style.sources.mapbox = {
@@ -420,108 +334,6 @@ describe('Map', () => {
             const map = createMap({cooperativeGestures: true});
             map.cooperativeGestures.disable();
             expect(map.getContainer().querySelector('.maplibregl-cooperative-gesture-screen')).toBeFalsy();
-        });
-    });
-
-    describe('#calculateCameraOptionsFromTo', () => {
-        // Choose initial zoom to avoid center being constrained by mercator latitude limits.
-        test('pitch 90 with terrain', () => {
-            const map = createMap();
-
-            const mockedGetElevation = jest.fn((_lngLat: LngLat, _zoom: number) => 111200);
-
-            const terrainStub = {} as Terrain;
-            terrainStub.getElevationForLngLatZoom = mockedGetElevation;
-            map.terrain = terrainStub;
-
-            // distance between lng x and lng x+1 is 111.2km at same lat
-            // altitude same as center elevation => 90° pitch
-            const cameraOptions: CameraOptions = map.calculateCameraOptionsFromTo(new LngLat(1, 0), 111200, new LngLat(0, 0));
-            expect(cameraOptions).toBeDefined();
-            expect(cameraOptions.pitch).toBeCloseTo(90);
-            expect(mockedGetElevation.mock.calls).toHaveLength(1);
-        });
-
-        test('pitch 153.435 with terrain', () => {
-            const map = createMap();
-
-            const mockedGetElevation = jest.fn((_lngLat: LngLat, _zoom: number) => 111200 * 3);
-
-            const terrainStub = {} as Terrain;
-            terrainStub.getElevationForLngLatZoom = mockedGetElevation;
-            map.terrain = terrainStub;
-            // distance between lng x and lng x+1 is 111.2km at same lat
-            // (elevation difference of cam and center) / 2 = grounddistance =>
-            // acos(111.2 / sqrt(111.2² + (111.2 * 2)²)) = acos(1/sqrt(5)) => 63.435 + 90 = 153.435
-            const cameraOptions: CameraOptions = map.calculateCameraOptionsFromTo(new LngLat(1, 0), 111200, new LngLat(0, 0));
-            expect(cameraOptions).toBeDefined();
-            expect(cameraOptions.pitch).toBeCloseTo(153.435);
-            expect(mockedGetElevation.mock.calls).toHaveLength(1);
-        });
-
-        test('pitch 63 with terrain', () => {
-            const map = createMap();
-
-            const mockedGetElevation = jest.fn((_lngLat: LngLat, _zoom: number) => 111200 / 2);
-
-            const terrainStub = {} as Terrain;
-            terrainStub.getElevationForLngLatZoom = mockedGetElevation;
-            map.terrain = terrainStub;
-
-            // distance between lng x and lng x+1 is 111.2km at same lat
-            // (elevation difference of cam and center) * 2 = grounddistance =>
-            // acos(111.2 / sqrt(111.2² + (111.2 * 0.5)²)) = acos(1/sqrt(1.25)) => 90 (looking down) - 26.565 = 63.435
-            const cameraOptions: CameraOptions = map.calculateCameraOptionsFromTo(new LngLat(0, 0), 111200, new LngLat(1, 0));
-            expect(cameraOptions).toBeDefined();
-            expect(cameraOptions.pitch).toBeCloseTo(63.435);
-            expect(mockedGetElevation.mock.calls).toHaveLength(1);
-        });
-
-        test('zoom distance 1000', () => {
-            const map = createMap();
-
-            const mockedGetElevation = jest.fn((_lngLat: LngLat, _zoom: number) => 1000);
-
-            const terrainStub = {} as Terrain;
-            terrainStub.getElevationForLngLatZoom = mockedGetElevation;
-            map.terrain = terrainStub;
-
-            const expectedZoom = Math.log2(map.transform.cameraToCenterDistance / mercatorZfromAltitude(1000, 0) / map.transform.tileSize);
-            const cameraOptions = map.calculateCameraOptionsFromTo(new LngLat(0, 0), 0, new LngLat(0, 0));
-
-            expect(cameraOptions).toBeDefined();
-            expect(cameraOptions.zoom).toBeCloseTo(expectedZoom);
-            expect(mockedGetElevation.mock.calls).toHaveLength(1);
-        });
-
-        test('don\'t call getElevation when altitude supplied', () => {
-            const map = createMap();
-
-            const mockedGetElevation = jest.fn((_tileID: OverscaledTileID, _x: number, _y: number, _extent?: number) => 0);
-
-            const terrainStub = {} as Terrain;
-            terrainStub.getElevation = mockedGetElevation;
-            map.terrain = terrainStub;
-
-            const cameraOptions = map.calculateCameraOptionsFromTo(new LngLat(0, 0), 0, new LngLat(0, 0), 1000);
-
-            expect(cameraOptions).toBeDefined();
-            expect(mockedGetElevation.mock.calls).toHaveLength(0);
-        });
-
-        test('don\'t call getElevation when altitude 0 supplied', () => {
-            const map = createMap();
-
-            const mockedGetElevation = jest.fn((_tileID: OverscaledTileID, _x: number, _y: number, _extent?: number) => 0);
-
-            const terrainStub = {} as Terrain;
-            terrainStub.getElevation = mockedGetElevation;
-            map.terrain = terrainStub;
-
-            const cameraOptions = map.calculateCameraOptionsFromTo(new LngLat(0, 0), 0, new LngLat(1, 0), 0);
-
-            expect(cameraOptions).toBeDefined();
-            expect(mockedGetElevation.mock.calls).toHaveLength(0);
         });
     });
 });
