+++ src/ui/map_feature_state.test.ts
