+++ src/ui/map_world_copies.test.ts
